lms-ai-agent/backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.23.2
sqlalchemy==1.4.49
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
langchain-community==0.4.1
langchain-openai==1.0.3